split LEM circuit code to factor out MultiFrame implementations
